AND-3811 Moved UserWalletId model to :domain:models module
