+++ app/src/main/java/com/tangem/tap/domain/model/UserWallet.kt
@@ -1,7 +1,7 @@
 package com.tangem.tap.domain.model
 
 import com.tangem.domain.models.scan.ScanResponse
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 
 /**
  * Represents user's wallet which stored in app persistence
