+++ app/src/main/java/com/tangem/tap/domain/model/WalletStoreModel.kt
@@ -3,7 +3,7 @@ package com.tangem.tap.domain.model
 import com.tangem.blockchain.common.Blockchain
 import com.tangem.blockchain.common.WalletManager
 import com.tangem.crypto.hdWallet.DerivationPath
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.domain.model.WalletStoreModel.WalletRent
 import com.tangem.tap.domain.tokens.models.BlockchainNetwork
 import com.tangem.tap.features.wallet.models.Currency
