+++ app/src/main/java/com/tangem/tap/domain/userWalletList/utils/Mapper.kt
@@ -1,6 +1,6 @@
 package com.tangem.tap.domain.userWalletList.utils
 
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.domain.model.UserWallet
 import com.tangem.tap.domain.userWalletList.model.UserWalletPublicInformation
 import com.tangem.tap.domain.userWalletList.model.UserWalletSensitiveInformation
