+++ app/src/main/java/com/tangem/tap/domain/walletStores/repository/WalletManagersRepository.kt
@@ -3,7 +3,7 @@ package com.tangem.tap.domain.walletStores.repository
 import com.tangem.blockchain.common.Blockchain
 import com.tangem.blockchain.common.WalletManager
 import com.tangem.common.CompletionResult
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.domain.model.UserWallet
 import com.tangem.tap.domain.tokens.models.BlockchainNetwork
