+++ app/src/main/java/com/tangem/tap/domain/walletStores/repository/WalletStoresRepository.kt
@@ -1,7 +1,7 @@
 package com.tangem.tap.domain.walletStores.repository
 
 import com.tangem.common.CompletionResult
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.domain.model.WalletStoreModel
 import com.tangem.tap.features.wallet.models.Currency
 import kotlinx.coroutines.flow.Flow
