+++ app/src/main/java/com/tangem/tap/domain/walletStores/repository/implementation/DefaultWalletStoresRepository.kt
@@ -2,7 +2,7 @@ package com.tangem.tap.domain.walletStores.repository.implementation
 
 import com.tangem.common.CompletionResult
 import com.tangem.common.catching
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.domain.model.WalletStoreModel
 import com.tangem.tap.domain.walletStores.repository.WalletStoresRepository
 import com.tangem.tap.domain.walletStores.repository.implementation.utils.isSameWalletStore
