+++ app/src/main/java/com/tangem/tap/features/walletSelector/redux/WalletSelectorMiddleware.kt
@@ -3,8 +3,8 @@ package com.tangem.tap.features.walletSelector.redux
 import com.tangem.common.*
 import com.tangem.common.core.TangemSdkError
 import com.tangem.core.analytics.Analytics
-import com.tangem.domain.common.util.UserWalletId
 import com.tangem.domain.models.scan.ScanResponse
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.*
 import com.tangem.tap.common.analytics.events.AnalyticsParam
 import com.tangem.tap.common.analytics.events.Basic
