+++ app/src/main/java/com/tangem/tap/features/walletSelector/redux/WalletSelectorState.kt
@@ -1,7 +1,7 @@
 package com.tangem.tap.features.walletSelector.redux
 
 import com.tangem.common.core.TangemError
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.common.entities.FiatCurrency
 import org.rekotlin.StateType
