+++ app/src/main/java/com/tangem/tap/features/walletSelector/ui/WalletSelectorScreenState.kt
@@ -1,7 +1,7 @@
 package com.tangem.tap.features.walletSelector.ui
 
 import androidx.compose.runtime.Immutable
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.features.details.ui.cardsettings.TextReference
 import com.tangem.tap.features.walletSelector.ui.model.DialogModel
 import com.tangem.tap.features.walletSelector.ui.model.MultiCurrencyUserWalletItem
