+++ app/src/main/java/com/tangem/tap/features/walletSelector/ui/components/MockData.kt
@@ -1,6 +1,6 @@
 package com.tangem.tap.features.walletSelector.ui.components
 
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.features.walletSelector.ui.WalletSelectorScreenState
 import com.tangem.tap.features.walletSelector.ui.model.MultiCurrencyUserWalletItem
 import com.tangem.tap.features.walletSelector.ui.model.SingleCurrencyUserWalletItem
@@ -8,7 +8,7 @@ import com.tangem.tap.features.walletSelector.ui.model.UserWalletItem
 
 internal object MockData {
     private val multiCurrencyUserWallet = MultiCurrencyUserWalletItem(
-        id = UserWalletId("wallet_1"),
+        id = UserWalletId.mock(n = 1),
         balance = UserWalletItem.Balance.Loaded(
             amount = "6781.05 $",
             showWarning = true,
@@ -21,7 +21,7 @@ internal object MockData {
     )
 
     private val singleCurrencyUserWallet = SingleCurrencyUserWalletItem(
-        id = UserWalletId("wallet_4"),
+        id = UserWalletId.mock(n = 4),
         balance = UserWalletItem.Balance.Loaded(
             amount = "6781.05 $",
             showWarning = false,
@@ -35,8 +35,8 @@ internal object MockData {
     val state = WalletSelectorScreenState(
         multiCurrencyWallets = listOf(
             multiCurrencyUserWallet,
-            multiCurrencyUserWallet.copy(id = UserWalletId("wallet_2")),
-            multiCurrencyUserWallet.copy(id = UserWalletId("wallet_3"), tokensCount = 2, cardsInWallet = 1),
+            multiCurrencyUserWallet.copy(id = UserWalletId.mock(n = 2)),
+            multiCurrencyUserWallet.copy(id = UserWalletId.mock(n = 4), tokensCount = 2, cardsInWallet = 1),
         ),
         singleCurrencyWallets = listOf(singleCurrencyUserWallet),
         selectedUserWalletId = multiCurrencyUserWallet.id,
