+++ app/src/main/java/com/tangem/tap/features/walletSelector/ui/components/WalletItem.kt
@@ -4,16 +4,7 @@ import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.Divider
 import androidx.compose.material.Icon
@@ -39,7 +30,7 @@ import com.tangem.core.ui.components.SpacerH2
 import com.tangem.core.ui.components.SpacerW6
 import com.tangem.core.ui.components.SpacerW8
 import com.tangem.core.ui.res.TangemTheme
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.common.extensions.cardImageData
 import com.tangem.tap.features.walletSelector.ui.model.MultiCurrencyUserWalletItem
 import com.tangem.tap.features.walletSelector.ui.model.SingleCurrencyUserWalletItem
@@ -362,7 +353,7 @@ private fun WalletItemSample(modifier: Modifier = Modifier) {
     ) {
         WalletItem(
             wallet = MultiCurrencyUserWalletItem(
-                id = UserWalletId(value = null),
+                id = UserWalletId.mock(),
                 name = "Tangem Card",
                 imageUrl = "",
                 balance = UserWalletItem.Balance.Loaded(
@@ -381,7 +372,7 @@ private fun WalletItemSample(modifier: Modifier = Modifier) {
         Divider(modifier = Modifier.padding(vertical = 4.dp))
         WalletItem(
             wallet = MultiCurrencyUserWalletItem(
-                id = UserWalletId(value = null),
+                id = UserWalletId.mock(),
                 name = "Tangem Card",
                 imageUrl = "",
                 balance = UserWalletItem.Balance.Loaded(
@@ -400,7 +391,7 @@ private fun WalletItemSample(modifier: Modifier = Modifier) {
         Divider(modifier = Modifier.padding(vertical = 4.dp))
         WalletItem(
             wallet = MultiCurrencyUserWalletItem(
-                id = UserWalletId(value = null),
+                id = UserWalletId.mock(),
                 name = "Tangem Card",
                 imageUrl = "",
                 balance = UserWalletItem.Balance.Loading,
@@ -416,7 +407,7 @@ private fun WalletItemSample(modifier: Modifier = Modifier) {
         Divider(modifier = Modifier.padding(vertical = 4.dp))
         WalletItem(
             wallet = MultiCurrencyUserWalletItem(
-                id = UserWalletId(value = null),
+                id = UserWalletId.mock(),
                 name = "Tangem Card",
                 imageUrl = "",
                 balance = UserWalletItem.Balance.Failed,
@@ -432,7 +423,7 @@ private fun WalletItemSample(modifier: Modifier = Modifier) {
         Divider(modifier = Modifier.padding(vertical = 4.dp))
         WalletItem(
             wallet = SingleCurrencyUserWalletItem(
-                id = UserWalletId(value = null),
+                id = UserWalletId.mock(),
                 name = "Tangem Card",
                 imageUrl = "",
                 balance = UserWalletItem.Balance.Loaded(
