+++ app/src/main/java/com/tangem/tap/features/walletSelector/ui/model/UserWalletItem.kt
@@ -1,6 +1,6 @@
 package com.tangem.tap.features.walletSelector.ui.model
 
-import com.tangem.domain.common.util.UserWalletId
+import com.tangem.domain.models.userwallet.UserWalletId
 import com.tangem.tap.features.wallet.redux.utils.UNKNOWN_AMOUNT_SIGN
 
 internal sealed interface UserWalletItem {
